Split General tab into Connection and System tabs
